[Fixes: #160566381] Fixes the enpoints
