+++ app/tests/v1/test_views.py
@@ -1,4 +1,4 @@
-#app/tests/v1/test_views.py
+#app/tests/v1/test_auth.py
 
 '''Implement classes and methods for testing API Endpoints.'''
 
@@ -6,182 +6,172 @@ import json
 import unittest
 from app import create_app
 
-class TestRouteCases(unittest.TestCase):
-    '''This class represents the orders test cases'''
+
+class TestAuthentication(unittest.TestCase):
+    '''This class represents user authentication unit tests'''
+
 
     def setUp(self):
         '''Define variables and initialize app'''
         self.app = create_app(app_config_name="testing")
         self.client = self.app.test_client
-        self.order = {
-            'order_item': 'Tater tots',
-            'description': 'These are Pieces of deep-fried, potatoes.',
-            'quantity': 7,
-            'status': 'Accepted'
+        self.user = {
+            'username': 'Promaster',
+            'password': 'Promaster2018',
         }
 
     def test_config(self):
         """Test configurations"""
         self.assertEqual(self.app.testing, True)
 
-    def test_user_can_place_an_order_if_not_dublicate(self):
-        '''Test API can create a new order (POST request)'''
-        self.client().post(
-            '/api/v1/orders',
-            data=json.dumps(self.order),
-            headers={'content-type': 'application/json'}
-        )
+    def test_register_with_valid_inputs(self):
+        """User can register for a new account"""
         resp = self.client().post(
-            '/api/v1/orders',
-            data=json.dumps(self.order),
+            '/api/v1/register',
+            data=json.dumps(self.user),
             headers={'content-type': 'application/json'}
         )
         response = json.loads(resp.data.decode('utf-8'))
-        self.assertEqual(resp.status_code, 400, response["Message"])
+        self.assertEqual(resp.status_code, 201, response['Message'])
 
-    def test_attempt_to_place_orders_without_order_item(self):
-        '''Test user attempts to place Unknown order i.e. without the order item'''
-        self.order = {
-            'description': 'These are Pieces of deep-fried, potatoes.',
-            'quantity': 7,
-            'status': 'Accepted'
+    def test_register_with_empty_username(self):
+        """Test when username is left blank"""
+        newuser = {
+            "username": "",
+            "password": "Promaster2018"
         }
         resp = self.client().post(
-            '/api/v1/orders',
-            data=json.dumps(self.order),
+            '/api/v1/register',
+            data=json.dumps(newuser),
             headers={'content-type': 'application/json'}
         )
         response = json.loads(resp.data.decode('utf-8'))
-        self.assertEqual(resp.status_code, 400, response["Message"])
+        self.assertEqual(resp.status_code, 400, response['Message'])
 
-    def test_user_can_get_all_orders(self):
-        '''Test API can return all the orders (GET request)'''
-        resp = self.client().get(
-            '/api/v1/orders',
-            headers={'content-type': 'application/json'}
-        )
-        response = json.loads(resp.data.decode('utf-8'))
-        self.assertEqual(resp.status_code, 200, response["Message"])
-
-    def test_user_can_fetch_a_specific_order(self):
-        '''Test API can return a specific order using its id (GET request)'''
-        self.client().post(
-            '/api/v1/orders',
-            data=json.dumps(self.order),
-            headers={'content-type': 'application/json'}
-        )
-        resp = self.client().get(
-            '/api/v1/orders/1',
-            headers={'Content-type': 'application/json'}
-        )
-        response = json.loads(resp.data.decode('utf-8'))
-        self.assertEqual(resp.status_code, 200, response["Message"])
-
-    def test_user_attempt_to_fetch_un_existing_order(self):
-        '''Test cannot fetch an order that doesen't exist'''
-        self.client().post(
-            '/api/v1/orders',
-            data=json.dumps(self.order),
-            headers={'content-type': 'application/json'}
-        )
-        resp = self.client().get(
-            '/api/v1/orders/2',
-            headers={'Content-type': 'application/json'}
-        )
-        response = json.loads(resp.data.decode('utf-8'))
-        self.assertEqual(resp.status_code, 404, response["Message"])
-
-    def test_user_can_update_order_status(self):
-        '''Test API can modify the status of an order (PUT request)'''
-        self.client().post(
-            '/api/v1/orders',
-            data=json.dumps(self.order),
-            headers={'content-type': 'application/json'}
-        )
-        resp = self.client().put(
-            '/api/v1/orders/1',
-            data=json.dumps({'status': 'Accepted'}),
-            headers={'content-type': 'application/json'}
-        )
-        response = json.loads(resp.data.decode('utf-8'))
-        self.assertEqual(resp.status_code, 200, response["Message"])
-
-    def test_for_attempt_to_update_un_existing_order(self):
-        """Unexisting order cannot be updated"""
-        self.client().post(
-            '/api/v1/orders',
-            data=json.dumps(self.order),
-            headers={'content-type': 'application/json'}
-        )
-        resp = self.client().put(
-            '/api/v1/orders/2',
-            data=json.dumps({'status': 'Accepted'}),
-            headers={'content-type': 'application/json'}
-        )
-        response = json.loads(resp.data.decode('utf-8'))
-        self.assertEqual(resp.status_code, 404, response["Message"])
-
-    def test_user_can_delete_an_order(self):
-        '''Test API can delete an order (DELETE request)'''
-        self.client().post(
-            '/api/v1/orders',
-            data=json.dumps(self.order),
-            headers={'content-type': 'application/json'}
-        )
-        resp = self.client().delete(
-            'api/v1/orders/1',
-            headers={'Content-type': 'application/json'}
-        )
-        response = json.loads(resp.data.decode('utf-8'))
-        self.assertEqual(resp.status_code, 200, response["Message"])
-
-    def test_for_attempt_to_delete_un_existing_order(self):
-        """Unexisting order cannot be updated"""
-        self.client().post(
-            '/api/v1/orders',
-            data=json.dumps(self.order),
-            headers={'content-type': 'application/json'}
-        )
-        resp = self.client().delete(
-            'api/v1/orders/2',
-            headers={'Content-type': 'application/json'}
-        )
-        response = json.loads(resp.data.decode('utf-8'))
-        self.assertEqual(resp.status_code, 404, response["Message"])
-
-    def test_malformed_url_for_place_order(self):
-        '''Test API should return an error: Not found'''
+    def test_register_with_short_username(self):
+        """When username is less than 6 characters"""
+        newuser = {
+            "username": "paul",
+            "password": "Promaster2018"
+        }
         resp = self.client().post(
-            '/api/v1/orders/',
-            data=json.dumps(self.order),
+            '/api/v1/register',
+            data=json.dumps(newuser),
             headers={'content-type': 'application/json'}
         )
-        self.assertEqual(resp.status_code, 404)
+        response = json.loads(resp.data.decode('utf-8'))
+        self.assertEqual(resp.status_code, 400, response['Message'])
 
-    def test_malformed_url_for_get_all_orders(self):
-        '''Test API should return an error: Not found'''
+    def test_register_without_password(self):
+        """User tries to register without a password"""
+        newuser = {
+            "username": "Promaster",
+            "password": ""
+        }
+        resp = self.client().post(
+            '/api/v1/register',
+            data=json.dumps(newuser),
+            headers={'content-type': 'application/json'}
+        )
+        response = json.loads(resp.data.decode('utf-8'))
+        self.assertEqual(resp.status_code, 400, response['Message'])
+
+    def test_register_with_short_password(self):
+        """Password less than 8 alphanumeric characters"""
+        newuser = {
+            "username": "Promaster",
+            "password": "Pm2018"
+        }
+        resp = self.client().post(
+            '/api/v1/register',
+            data=json.dumps(newuser),
+            headers={'content-type': 'application/json'}
+        )
+        response = json.loads(resp.data.decode('utf-8'))
+        self.assertEqual(resp.status_code, 400, response['Message'])
+
+    def test_register_for_password_without_numbers(self):
+        """Password without atleast one numeric character"""
+        newuser = {
+            "username": "Promaster",
+            "password": "Promaster"
+        }
+        resp = self.client().post(
+            '/api/v1/register',
+            data=json.dumps(newuser),
+            headers={'content-type': 'application/json'}
+        )
+        response = json.loads(resp.data.decode('utf-8'))
+        self.assertEqual(resp.status_code, 400, response['Message'])
+
+    def test_register_for_password_without_capital_letters(self):
+        """Password without atleast one capital letter"""
+        newuser = {
+            "username": "Promaster",
+            "password": "promaster2018"
+        }
+        resp = self.client().post(
+            '/api/v1/register',
+            data=json.dumps(newuser),
+            headers={'content-type': 'application/json'}
+        )
+        response = json.loads(resp.data.decode('utf-8'))
+        self.assertEqual(resp.status_code, 400, response['Message'])
+
+    def test_login_with_unregistered_username(self):
+        """User enters unregistered username"""
+        newuser = {
+            "username": "Johnson",
+            "password": "promaster2018"
+        }
+        self.client().post(
+            '/api/v1/register',
+            data=json.dumps(newuser),
+            headers={'content-type': 'application/json'}
+        )
         resp = self.client().get(
-            '/api/v1/orders/',
+            '/api/v1/login/username/Johnson/password/Promaster2018',
             headers={'content-type': 'application/json'}
         )
-        self.assertEqual(resp.status_code, 404)
+        response = json.loads(resp.data.decode('utf-8'))
+        self.assertEqual(resp.status_code, 401, response['Message'])
 
-    def test_malformed_url_for_get_a_specific_order(self):
-        '''Test API should return an error: Not found'''
+    def test_login_with_wrong_password(self):
+        """Login with wrong password"""
+        newuser = {
+            "username": "Johnson",
+            "password": "promaster2018"
+        }
+        self.client().post(
+            '/api/v1/register',
+            data=json.dumps(newuser),
+            headers={'content-type': 'application/json'}
+        )
         resp = self.client().get(
-            '/api/v1/orders/-1',
+            '/api/v1/login/username/Johnson/password/Pomaster2018',
             headers={'content-type': 'application/json'}
         )
-        self.assertEqual(resp.status_code, 404)
+        response = json.loads(resp.data.decode('utf-8'))
+        self.assertEqual(resp.status_code, 401, response['Message'])
 
-    def test_malformed_url_for_update_an_order(self):
-        '''Test API should return an error: Not found'''
-        resp = self.client().put(
-            '/api/v1/orders/-1',
-            data=json.dumps({'status': 'Accepted'}),
+    def test_login_with_valid_credentials(self):
+        """Login with valid username and password"""
+        newuser = {
+            "username": "Johnson",
+            "password": "Promaster2018"
+        }
+        self.client().post(
+            '/api/v1/register',
+            data=json.dumps(newuser),
             headers={'content-type': 'application/json'}
         )
-        self.assertEqual(resp.status_code, 404)
-# 
-# if __name__ == "__main__":
-#     unittest.main()
+        resp = self.client().get(
+            '/api/v1/login/username/Johnson/password/Promaster2018',
+            headers={'content-type': 'application/json'}
+        )
+        response = json.loads(resp.data.decode('utf-8'))
+        self.assertEqual(resp.status_code, 200, response['Message'])
+
+if __name__ == "__main__":
+    unittest.main()
+
